time updata + norm

+++ philo/src/verif/convert.c
@@ -6,7 +6,7 @@
 /*   By: besalort <besalort@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/23 15:00:48 by besalort          #+#    #+#             */
-/*   Updated: 2023/10/23 16:24:56 by besalort         ###   ########.fr       */
+/*   Updated: 2023/10/23 16:54:50 by besalort         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,15 +17,18 @@ void	convert_all(t_data *data, char **av)
 	convert_die(data, av[2]);
 	convert_eat(data, av[3]);
 	convert_sleep(data, av[4]);
-	printf("%06ld est le temps en microseconde pour mourir\n", data->time_die.tv_usec);
-	printf("%06ld est le temps en microseconde pour manger\n", data->time_eat.tv_usec);
-	printf("%06ld est le temps en microseconde pour dormir\n", data->time_sleep.tv_usec);
+	printf("%06ld est le temps en microseconde pour mourir\n",
+		data->time_die.tv_usec);
+	printf("%06ld est le temps en microseconde pour manger\n",
+		data->time_eat.tv_usec);
+	printf("%06ld est le temps en microseconde pour dormir\n",
+		data->time_sleep.tv_usec);
 }
 
 void	convert_die(t_data *data, char *time)
 {
 	int				i;
-	int 			count;
+	long int		count;
 
 	i = 0;
 	time = invert_tab(time);
@@ -41,11 +44,10 @@ void	convert_die(t_data *data, char *time)
 	data->time_die.tv_usec = count;
 }
 
-
 void	convert_eat(t_data *data, char *time)
 {
-	int				i;
-	int 			count;
+	int			i;
+	long int	count;
 
 	i = 0;
 	time = invert_tab(time);
@@ -63,8 +65,8 @@ void	convert_eat(t_data *data, char *time)
 
 void	convert_sleep(t_data *data, char *time)
 {
-	int						i;
-	long long int 			count;
+	int					i;
+	long int			count;
 
 	i = 0;
 	time = invert_tab(time);
@@ -88,7 +90,7 @@ char	*invert_tab(char *tab)
 
 	i = 0;
 	j = 0;
-	while (tab && tab[j+1])
+	while (tab && tab[j + 1])
 		j++;
 	while (i < j)
 	{
